+++ src/main/java/page/clab/api/global/common/slack/domain/AlertType.java
@@ -1,0 +1,9 @@
+package page.clab.api.global.common.slack.domain;
+
+public interface AlertType {
+
+    String getTitle();
+
+    String getDefaultMessage();
+
+}
