+++ src/main/java/page/clab/api/global/common/slack/domain/SecurityAlertType.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum SecurityAlertType {
+public enum SecurityAlertType implements AlertType {
 
     ABNORMAL_ACCESS("비정상적인 접근", "Unexpected access pattern detected."),
     REPEATED_LOGIN_FAILURES("지속된 로그인 실패", "Multiple consecutive failed login attempts."),
@@ -20,5 +20,4 @@ public enum SecurityAlertType {
 
     private final String title;
     private final String defaultMessage;
-
 }
